refactor: Move python lib into separate directory
